requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# OAuth and HTTP
httpx==0.26.0
authlib==1.3.0
itsdangerous==2.1.2

# Google Calendar
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0

# Microsoft Graph
msal==1.26.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.1
pytz==2024.1
